Cleanup router module
